refactor ActionBuilder
