+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/planner/goap/ActionBuilder.kt
@@ -5,12 +5,6 @@ package ai.koog.agents.planner.goap
 /**
  * Builder for [Action] instances.
  */
-public expect class ActionBuilder<State>() : ActionBuilderApi<State> {
-    public override fun name(name: String): ActionBuilder<State>
-    public override fun description(description: String?): ActionBuilder<State>
-    public override fun precondition(precondition: Condition<State>): ActionBuilder<State>
-    public override fun belief(belief: Belief<State>): ActionBuilder<State>
-    public override fun cost(cost: Cost<State>): ActionBuilder<State>
-    public override fun execute(execute: Execute<State>): ActionBuilder<State>
-    public override fun build(): Action<State>
+public expect class ActionBuilder<State>() : ActionBuilderCommon<State, ActionBuilder<State>> {
+    public override fun self(): ActionBuilder<State>
 }
